# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CheckInRequest:
    properties:
      location:
        example: Jakarta
        minLength: 3
        type: string
    required:
    - location
    type: object
  dto.CheckInResponse:
    properties:
      check_in:
        example: "2022-01-01T00:00:00Z"
        type: string
      date:
        example: "2022-01-01"
        type: string
      location:
        example: Jakarta
        type: string
      status:
        example: present
        type: string
      user_id:
        example: 00000000-0000-0000-0000-000000000000
        type: string
    type: object
...
  types.ErrorResponse:
    properties:
      error: {}
      message:
        example: Unauthorized
        type: string
      status:
        example: 401
        type: integer
    type: object
  types.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
host: localhost:5000
info:
  contact:
    email: ilham15suandi@gmail.com
    name: Ilham Suandi
    url: https://github.com/ilhamSuandi
  description: Growy API
  title: Growy API
  version: 1.0.0
paths:
...
  /attendances/check-in/{token}:
    post:
      consumes:
      - application/json
      description: daily checkin user, use qrcode to check in
      parameters:
      - description: token
        in: path
        name: token
        required: true
        type: string
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CheckInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully Checked In Attendance
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CheckInResponse'
              type: object
        "400":
          description: request body is invalid
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: token is empty
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: check in user
      tags:
      - Attendance
...
securityDefinitions:
  BearerAuth:
    description: 'Example Value: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"